+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add custom margins to pinToSuperViewEdges
+
 ## [0.0.2] - 2021-02-10
 
 ### Added